+++ bitframe-utils/presenters/core/src/commonMain/kotlin/presenters/collections/table/RowActionsBuilder.kt
@@ -1,4 +1,4 @@
-package bitframe.presenters.collections.table
+package presenters.collections.table
 
 import kotlinx.collections.interoperable.mutableListOf