+++ bitframe-utils/presenters/core/src/commonMain/kotlin/presenters/feedbacks/FormFeedback.kt
@@ -1,6 +1,6 @@
 @file:JsExport
 
-package bitframe.presenters.feedbacks
+package presenters.feedbacks
 
 import kotlin.js.JsExport