+++ pi-monitor/pi-monitor-client/browser/react/src/jsMain/kotlin/pimonitor/authentication/signup/SignUp.kt
@@ -1,9 +1,8 @@
 package pimonitor.authentication.signup
 
 import bitframe.PanelPageRoute
-import pimonitor.authentication.signup.SignUpState
 import bitframe.components.TextInput
-import bitframe.presenters.feedbacks.FormFeedback
+import presenters.feedbacks.FormFeedback
 import kotlinx.css.*
 import kotlinx.extensions.onDesktop
 import kotlinx.extensions.onMobile
@@ -19,7 +18,6 @@ import react.router.dom.useHistory
 import react.router.dom.withRouter
 import reakt.*
 import pimonitor.reakt.DropDown
-import pimonitor.authentication.signup.SignUpParams
 import styled.css
 import styled.styledH2
 import theme.clazz